Clone request in tests
This clone is ESSENTIAL. Symfony2 stores the matched controller inside
the request. Since we are replaying the requests from the first test in
the second tests and this uses the same request object, the route would
not be matched again. This causes the cotnroller from the first test run
being re-used, which also means that the first few requests run against
the storage from the first run.